refactor(github-tool): enhance pull request data fetching and type safety
- Update FileMetadata type to allow null values for optional fields
- Introduce new fetchers for pull request data retrieval, improving modularity
- Consolidate fetching logic for pull request info and diffs into dedicated functions
- Remove redundant code from loader.ts, streamlining the implementation
